Update app.js
uncommented the app.post to test.  post should work like aismarthome but i noticed that the app.post line between aismarthome and myjsonparser is different so when combining them, it might fail because the post aismarthome receives it to api/switches whereas the post myjsonparser handles is to root '/'.

Added random function but commented it.